finalprojectw23-studywhere-main/backend/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
mysqlclient==2.1.1
protobuf==3.20.3
PyMySQL==1.0.2
pyodbc==4.0.35
PyYAML==6.0
requests==2.28.2
SQLAlchemy==2.0.2
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
